add ability to test for dialect transaction support

+++ test/associations/has-one.test.js
@@ -23,28 +23,30 @@ describe(Support.getTestDialectTeaser("HasOne"), function() {
   })
 
   describe('getAssocation', function() {
-    it('supports transactions', function(done) {
-      Support.prepareTransactionTest(this.sequelize, function(sequelize) {
-        var User  = sequelize.define('User', { username: Support.Sequelize.STRING })
-          , Group = sequelize.define('Group', { name: Support.Sequelize.STRING })
+    if (current.dialect.supports.transactions) {
+      it('supports transactions', function(done) {
+        Support.prepareTransactionTest(this.sequelize, function(sequelize) {
+          var User  = sequelize.define('User', { username: Support.Sequelize.STRING })
+            , Group = sequelize.define('Group', { name: Support.Sequelize.STRING })
 
-        Group.hasOne(User)
+          Group.hasOne(User)
 
-        sequelize.sync({ force: true }).success(function() {
-          User.create({ username: 'foo' }).success(function(fakeUser) {
-            User.create({ username: 'foo' }).success(function(user) {
-              Group.create({ name: 'bar' }).success(function(group) {
-                sequelize.transaction().then(function(t) {
-                  group.setUser(user, { transaction: t }).success(function() {
-                    Group.all().success(function(groups) {
-                      groups[0].getUser().success(function(associatedUser) {
-                        expect(associatedUser).to.be.null
-                        Group.all({ transaction: t }).success(function(groups) {
-                          groups[0].getUser({ transaction: t }).success(function(associatedUser) {
-                            expect(associatedUser).not.to.be.null
-                            expect(associatedUser.id).to.equal(user.id)
-                            expect(associatedUser.id).not.to.equal(fakeUser.id)
-                            t.rollback().success(function() { done() })
+          sequelize.sync({ force: true }).success(function() {
+            User.create({ username: 'foo' }).success(function(fakeUser) {
+              User.create({ username: 'foo' }).success(function(user) {
+                Group.create({ name: 'bar' }).success(function(group) {
+                  sequelize.transaction().then(function(t) {
+                    group.setUser(user, { transaction: t }).success(function() {
+                      Group.all().success(function(groups) {
+                        groups[0].getUser().success(function(associatedUser) {
+                          expect(associatedUser).to.be.null
+                          Group.all({ transaction: t }).success(function(groups) {
+                            groups[0].getUser({ transaction: t }).success(function(associatedUser) {
+                              expect(associatedUser).not.to.be.null
+                              expect(associatedUser.id).to.equal(user.id)
+                              expect(associatedUser.id).not.to.equal(fakeUser.id)
+                              t.rollback().success(function() { done() })
+                            })
                           })
                         })
                       })
@@ -56,7 +58,7 @@ describe(Support.getTestDialectTeaser("HasOne"), function() {
           })
         })
       })
-    })
+    }
 
     it('does not modify the passed arguments', function () {
       var User = this.sequelize.define('user', {})
@@ -98,36 +100,38 @@ describe(Support.getTestDialectTeaser("HasOne"), function() {
   })
 
   describe('setAssociation', function() {
-    it('supports transactions', function(done) {
-      Support.prepareTransactionTest(this.sequelize, function(sequelize) {
-        var User  = sequelize.define('User', { username: Support.Sequelize.STRING })
-          , Group = sequelize.define('Group', { name: Support.Sequelize.STRING })
+    if (current.dialect.supports.transactions) {
+      it('supports transactions', function(done) {
+        Support.prepareTransactionTest(this.sequelize, function(sequelize) {
+          var User  = sequelize.define('User', { username: Support.Sequelize.STRING })
+            , Group = sequelize.define('Group', { name: Support.Sequelize.STRING })
 
-        Group.hasOne(User)
+          Group.hasOne(User)
 
-        sequelize.sync({ force: true }).success(function() {
-          User.create({ username: 'foo' }).success(function(user) {
-            Group.create({ name: 'bar' }).success(function(group) {
-              sequelize.transaction().then(function(t) {
-                group
-                  .setUser(user, { transaction: t })
-                  .success(function() {
-                    Group.all().success(function(groups) {
-                      groups[0].getUser().success(function(associatedUser) {
-                        expect(associatedUser).to.be.null
-                        t.rollback().success(function() { done() })
+          sequelize.sync({ force: true }).success(function() {
+            User.create({ username: 'foo' }).success(function(user) {
+              Group.create({ name: 'bar' }).success(function(group) {
+                sequelize.transaction().then(function(t) {
+                  group
+                    .setUser(user, { transaction: t })
+                    .success(function() {
+                      Group.all().success(function(groups) {
+                        groups[0].getUser().success(function(associatedUser) {
+                          expect(associatedUser).to.be.null
+                          t.rollback().success(function() { done() })
+                        })
                       })
                     })
-                  })
-                  .on('sql', function(sql, uuid) {
-                    expect(uuid).to.not.equal('default')
-                  })
+                    .on('sql', function(sql, uuid) {
+                      expect(uuid).to.not.equal('default')
+                    })
+                })
               })
             })
           })
         })
       })
-    })
+    }
 
     it('can set an association with predefined primary keys', function(done) {
       var User = this.sequelize.define('UserXYZZ', { userCoolIdTag: { type: Sequelize.INTEGER, primaryKey: true }, username: Sequelize.STRING })
@@ -227,24 +231,26 @@ describe(Support.getTestDialectTeaser("HasOne"), function() {
       })
     })
 
-    it('supports transactions', function(done) {
-      Support.prepareTransactionTest(this.sequelize, function(sequelize) {
-        var User  = sequelize.define('User', { username: Sequelize.STRING })
-          , Group = sequelize.define('Group', { name: Sequelize.STRING })
+    if (current.dialect.supports.transactions) {
+      it('supports transactions', function(done) {
+        Support.prepareTransactionTest(this.sequelize, function(sequelize) {
+          var User  = sequelize.define('User', { username: Sequelize.STRING })
+            , Group = sequelize.define('Group', { name: Sequelize.STRING })
 
-        User.hasOne(Group)
+          User.hasOne(Group)
 
-        sequelize.sync({ force: true }).success(function() {
-          User.create({ username: 'bob' }).success(function(user) {
-            sequelize.transaction().then(function(t) {
-              user.createGroup({ name: 'testgroup' }, { transaction: t }).success(function() {
-                User.all().success(function (users) {
-                  users[0].getGroup().success(function (group) {
-                    expect(group).to.be.null;
-                    User.all({ transaction: t }).success(function (users) {
-                      users[0].getGroup({ transaction: t }).success(function (group) {
-                        expect(group).to.be.not.null;
-                        t.rollback().success(function() { done() })
+          sequelize.sync({ force: true }).success(function() {
+            User.create({ username: 'bob' }).success(function(user) {
+              sequelize.transaction().then(function(t) {
+                user.createGroup({ name: 'testgroup' }, { transaction: t }).success(function() {
+                  User.all().success(function (users) {
+                    users[0].getGroup().success(function (group) {
+                      expect(group).to.be.null;
+                      User.all({ transaction: t }).success(function (users) {
+                        users[0].getGroup({ transaction: t }).success(function (group) {
+                          expect(group).to.be.not.null;
+                          t.rollback().success(function() { done() })
+                        })
                       })
                     })
                   })
@@ -254,7 +260,8 @@ describe(Support.getTestDialectTeaser("HasOne"), function() {
           })
         })
       })
-    })
+    }
+
   })
 
   describe('foreign key', function () {
